# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NWMS API",
        "description": "This API uses Bearer Token authentication.\n *     To authorize, use the 'Authorize' button above and enter a token in the format:\n *     Bearer {your_token}",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "operationId": "78ce1bd22726e99d8f30c6c8fa329168",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create new products",
                "operationId": "1e5b76a5975af81b35765edfe66f34ec",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Product"
                },
                "responses": {
                    "201": {
                        "description": "Products created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "operationId": "f4e69191da6baa0f13eb46ff6dd761ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "6f191ad01c85b6853580d01bd9c285cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Product"
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "b883dde0dd648a32e73f7b73699a2a85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Product": {
                "required": [
                    "of_name",
                    "of_shop_id"
                ],
                "properties": {
                    "of_ext_id": {
                        "description": "External product ID from external system",
                        "type": "string",
                        "maxLength": 25
                    },
                    "of_shop_id": {
                        "description": "Shop ID",
                        "type": "integer"
                    },
                    "of_name": {
                        "description": "Product name",
                        "type": "string",
                        "maxLength": 150
                    },
                    "of_article": {
                        "description": "Product article number",
                        "type": "string",
                        "maxLength": 25
                    },
                    "of_sku": {
                        "description": "SKU code",
                        "type": "string",
                        "maxLength": 25
                    },
                    "of_price": {
                        "description": "Product base price",
                        "type": "number",
                        "format": "float"
                    },
                    "of_estimated_price": {
                        "description": "Estimated product price",
                        "type": "number",
                        "format": "float"
                    },
                    "of_img": {
                        "description": "Image URL",
                        "type": "string",
                        "maxLength": 255
                    },
                    "of_dimension_x": {
                        "description": "Length (cm)",
                        "type": "number",
                        "format": "float"
                    },
                    "of_dimension_y": {
                        "description": "Width (cm)",
                        "type": "number",
                        "format": "float"
                    },
                    "of_dimension_z": {
                        "description": "Height (cm)",
                        "type": "number",
                        "format": "float"
                    },
                    "of_weight": {
                        "description": "Weight (grams)",
                        "type": "integer"
                    },
                    "of_datamatrix": {
                        "description": "Requires DataMatrix (0 or 1)",
                        "type": "integer"
                    },
                    "of_comment": {
                        "description": "Comment or note",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
            "Product": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}